Two-Way Communication for setting the DSMX Pulse Rate on the ATT
New quadrotor needs a frame rate of 11 ms (default is 22 ms). So the byte frame sent to the Arduino was extended by a message type byte. The first frame after initializing the Arduino is the configuration, currently consisting only of the pulse rate in milliseconds. This change affects both the Arduino code and the ICARUS project.

Further preparations for setting the servo travel ways.

General cleanup, refactoring, orthography.